move simulator into lib
